requirements-dev.txt
appdirs==1.4.3
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.10.6
packaging==16.8
python-decouple==3.0
six==1.10.0
static3==0.7.0
djangorestframework
Pillow
django-storages
boto3
django-allauth
django-widget-tweaks
django-localflavor
django-filter
